refactor: moving SectionsMappingContext Axios requests into templateMapping service
